admin 추가 / admin header 추가

+++ Bookdream/src/main/java/com/spring/bookdream/controller/PayController.java
@@ -6,7 +6,7 @@ import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.util.List;
+
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -14,7 +14,9 @@ import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.bind.annotation.SessionAttributes;
 
@@ -29,7 +31,7 @@ import com.spring.bookdream.vo.PayVO;
 import com.spring.bookdream.vo.PurchaseVO;
 
 @Controller
-@SessionAttributes("payData")
+@SessionAttributes({"order", "payData"})
 @RequestMapping("/detail/cart/orderitem")
 public class PayController {
 	
@@ -47,6 +49,16 @@ public class PayController {
 	
 	@Autowired
 	private HttpSession session;	
+	
+	
+	// 결제 전 유저정보 세션등록
+	@RequestMapping(value="/save")
+	@ResponseBody
+	public void saveOrder(@RequestBody OrderVO order, Model model) {	
+		
+		model.addAttribute("order", order);
+
+	}
 	
 	// 결제 후 처리
 	@RequestMapping(value="/pay")
@@ -81,17 +93,26 @@ public class PayController {
 			
 		// 세션 호출 - 결제 후 DB에 넣는 목적
 		String pay_method = order.getPay_method();
+		int total_price = order.getTotal_price();
+		int final_price = order.getFinal_price();
 		int discount_price = order.getDiscount_price();
-		int final_price = order.getTotal_price();
-		int save_point = order.getSave_point();
+		int pay_fee = order.getPay_fee();
 		int use_point = order.getUse_point();
+		int save_point = order.getSave_point();
+		
+		String order_address = order.getOrder_address();
 		
 		pay.setPay_method(pay_method);
-		pay.setDiscount_price(discount_price);
-		pay.setFinal_price(final_price);
-		pay.setSave_point(save_point);
+		pay.setTotal_price(total_price);
+		pay.setFinal_price(final_price);		
+		pay.setDiscount_price(discount_price);	
+		pay.setPay_fee(pay_fee);
 		pay.setUse_point(use_point);
+		pay.setSave_point(save_point);
+
+		purchase.setOrder_address(order_address);
 		
+		// 유저번호 호출
 		int user_no = (int) session.getAttribute("user_no");
 		
 		order.setUser_no(user_no);
@@ -105,36 +126,65 @@ public class PayController {
 	    // 결제번호, 결제시간 추출 -> 다른 DB 등록
 	    PayVO payData = payService.searchPay(pay);
 	    
-
 	    order.setPay_no(payData.getPay_no());
-	    order.setOrder_enroll(payData.getPay_date());
+	    order.setOrder_date(payData.getPay_date());
 	    purchase.setOrder_no(payData.getPay_no());	    
 	    orderitem.setPay_no(payData.getPay_no());	   
 	    
 	    System.out.println("---> 주문 DB 등록 <---");
 	    orderService.insertOrder(order);			
-
-	    System.out.println("---> 주문상세보기 DB 등록 <---");
-	    PurchaseService.insertPurchase(purchase);
 	    
 	    System.out.println("---> 사용자 포인트 갱신 <---");
 	    orderitemService.updateUserPoint(orderitem);
 	   
-	    
-	    System.out.println("---> 구입한 개수만큼 재고차감 <---");
-	    orderitemService.updateBookStock(orderitem);
-	    
-	    // 장바구니에서 필터하는 기능이 없어서 지금은 전부다 제거됨 
-	    System.out.println("---> 구입한 상품만 장바구니 제거 <---");
-	    orderitemService.deleteCartList(orderitem);
+	    // 바로구매, 장바구니 판별
+	    String buy_now = (String) session.getAttribute("buy_now");
+
+	    // 바로구매
+	    if ("Y".equals(buy_now)) {
+
+		    int product_count = (int) session.getAttribute("product_count");	
+		    int book_no = (int) session.getAttribute("book_no");	
+		    
+	    	purchase.setProduct_count(product_count);
+	    	purchase.setBook_no(book_no);
+	    	
+		    System.out.println("---> 주문상세보기 DB 등록 (바로구매) <---");
+		    PurchaseService.insertPurchase_now(purchase);
+	    	
+	    	orderitem.setProduct_count(product_count);
+	    	orderitem.setBook_no(book_no);
+	    	
+	    	System.out.println("---> 바로구매 재고차감 <---");
+		    orderitemService.updateBookStock_now(orderitem);
+		    
+		// 장바구니 구매
+	    } else {
+	    	
+ 		    System.out.println("---> 주문상세보기 DB 등록 (장바구니) <---");
+		    PurchaseService.insertPurchase(purchase);
+		    
+		    System.out.println("---> 구입한 개수만큼 재고차감 (장바구니) <---");
+		    orderitemService.updateBookStock(orderitem);
+		    
+		    // 장바구니에서 필터하는 기능이 없어서 지금은 전부다 제거됨 
+		    System.out.println("---> 구입한 상품만 장바구니 제거 <---");
+		    orderitemService.deleteCartList(orderitem);	    	
+	    }
+
 	    
 	    // 결제번호, 결제시간 표시용
 	    model.addAttribute("payData", payData);
+	    
+	    session.removeAttribute("buy_now");
+	    session.removeAttribute("product_count");
+	    session.removeAttribute("book_no");
 	    
 	    return "redirect:/detail/cart/orderitem/success";
 	}		
 	
 		// 새로고침 DB중복 방지
+		// 방금 구매한 상품을 조회함
 		@RequestMapping(value="/success")
 		public String success(@SessionAttribute("payData") PayVO pay, PurchaseVO vo, Model model) {
 
@@ -143,7 +193,6 @@ public class PayController {
 			vo.setUser_no(user_no);
 			vo.setOrder_no(pay.getPay_no());			
 
-			
 			model.addAttribute("purchase", PurchaseService.getPurchaseList(vo));
 			
 			return "main/success";
